Fully implements user page functionality with Django backend
