+++ ios/Anywhere Reader/Anywhere Reader/Model/Article.swift
@@ -8,6 +8,37 @@
 
 import Foundation
 
+//typealias Articles = [Article]
+//
+//struct Article: Codable {
+//    let id: Int
+//    let title: String
+//    let author: String
+//    let normalURL: String
+//    let resolvedURL: String
+//    let dateSaved: String
+//    let datePublished: String
+//    let excerpt: String
+//    let coverImage: String
+//    let tags: String
+//    let text: String
+//
+//    enum CodingKeys: String, CodingKey {
+//        case id = "id"
+//        case title = "title"
+//        case author = "author"
+//        case normalURL = "normal_url"
+//        case resolvedURL = "resolved_url"
+//        case dateSaved = "date_saved"
+//        case datePublished = "date_published"
+//        case excerpt = "excerpt"
+//        case coverImage = "cover_image"
+//        case tags = "tags"
+//        case text = "text"
+//    }
+//}
+
+
 struct Articles: Codable {
     let articles: [Article]
 }
