+++ ios/Anywhere Reader/Anywhere Reader/View Controller/Content Collection/Collection View Cells/DocumentCollectionViewCell.swift
@@ -40,12 +40,5 @@ class DocumentCollectionViewCell: UICollectionViewCell {
         }
         
         titleLabel.text = article.title
-        
-//        let cornerRadius: CGFloat = 12.0
-//        shadowView.cornerRadius = cornerRadius
-//        blurLayer.layer.cornerRadius = cornerRadius
-//        imageView.layer.cornerRadius = cornerRadius
-        
-//        openButton.layer.cornerRadius = openButton.frame.height / 2.0
     }
 }
